chore: update gitignore to exclude Books and remove tracked binaries

+++ CSES/MaxSubArraySum.java
@@ -4,17 +4,17 @@ import java.util.*;
 public class MaxSubArraySum {
     public static void main(String[] args) {
         Scanner rd = new Scanner(System.in);
-        int n = rd.nextInt();
-        int[] arreglo = new int[n];
+        long n = rd.nextInt();
+        long[] arreglo = new long[(int) n];
         for (int i = 0; i < n; i++) {
             arreglo[i] = rd.nextInt();
         }
         System.out.println(maxSubArraySum(arreglo));
     }
 
-    public static int maxSubArraySum(int[] A) {
-        int max = Integer.MIN_VALUE;
-        int currentSum = 0;
+    public static long maxSubArraySum(long[] A) {
+        long max = Long.MIN_VALUE;
+        long currentSum = 0;
         for (int i = 0; i < A.length; i++) {
             if (i == 0) {
                 currentSum = A[i];
